// src/app/models/cliente/cliente.model.ts
import { TipoDocumento } from "./tipo-documento.type";

export interface ClienteModel {
  id: number | null;

  tipoDocumento: TipoDocumento | null;
  numeroDocumento: string | null;

  nombresApellidos: string;

  telefono: string | null;
  correo: string | null;
  direccion: string | null;
  observacion: string | null;

  fechaRegistro: string;
  activo: boolean;
}